+++ src/controllers/group-controller.js
@@ -1,6 +1,7 @@
 import axios from "axios";
 import { db } from "../models/db.js";
 import { LighthouseSpec} from "../models/joi-schemas.js";
+import { imageStore } from "../models/image-store.js";
 
 export const groupController = {
   index: {
@@ -33,6 +34,7 @@ export const groupController = {
         range: request.payload.range,
         latitude: request.payload.latitude,
         longitude: request.payload.longitude,
+        image: request.payload.image
       };
 
       const report = {};
@@ -67,6 +69,30 @@ export const groupController = {
       const group = await db.groupStore.getGroupById(request.params.id);
       await db.lighthouseStore.deleteLighthouse(request.params.lighthouseid);
       return h.redirect(`/group/${group._id}`);
+    },
+  },
+
+  uploadImage: {
+    handler: async function (request, h) {
+      try {
+        const group = await db.groupStore.getGroupById(request.params.id);
+        const file = request.payload.imagefile;
+        if (Object.keys(file).length > 0) {
+          const url = await imageStore.uploadImage(request.payload.imagefile);
+          group.img = url;
+          await db.groupStore.updateGroup(group);
+        }
+        return h.redirect(`/group/${group._id}`);
+      } catch (err) {
+        console.log(err);
+        return h.redirect(`/group/${group._id}`);
+      }
+    },
+    payload: {
+      multipart: true,
+      output: "data",
+      maxBytes: 209715200,
+      parse: true,
     },
   },
 };