+++ src/models/mongo/lighthouse-mongo-store.js
@@ -52,7 +52,6 @@ export const lighthouseMongoStore = {
     lighthouseDoc.range = updatedLighthouse.range;
     lighthouseDoc.latitude = updatedLighthouse.latitude;
     lighthouseDoc.longitude = updatedLighthouse.longitude;
-    lighthouseDoc.image = updatedLighthouse.image;
     
     await lighthouseDoc.save();
   },