+++ src/models/mongo/seed-data.js
@@ -24,6 +24,7 @@ export const seedData = {
       _model: "Group",
       myTestGroup: {
         title: "Test Group 1",
+        img: "http://res.cloudinary.com/dbx8kj5x4/image/upload/v1712088094/erxlob4tqjqne8wfdjr8.jpg",
         userid: "->users.bart"
       }
     },
@@ -38,7 +39,6 @@ export const seedData = {
       range: 23,
       latitude: 53.333,
       longitude: -7.444,
-      image: "",
         groupid: "->groups.myTestGroup"
       },
     }